--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cube"
-version = "0.5.2"
+version = "0.5.3"
 edition = "2021"
 
 [lib]