--- a/docs/index.html
+++ b/docs/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>cube solver</title>
-    <script type="module" crossorigin src="/cube/assets/index.c4dc8c1e.js"></script>
+    <script type="module" crossorigin src="/cube/assets/index.abbeb46c.js"></script>
     <link rel="stylesheet" href="/cube/assets/index.479570ec.css">
   </head>
   <body>