… Cargo.lock
… docs/assets/cube_bg.892becb5.wasm
… docs/assets/index.abbeb46c.js