Before:

```
   make benchmark
    Finished release [optimized] target(s) in 0.01s
Benchmark 1: cargo run --release -- --roux < tests/random/1
  Time (mean ± σ):     668.5 ms ±  29.6 ms    [User: 630.4 ms, System: 37.9 ms]
  Range (min … max):   636.2 ms … 724.6 ms    20 runs

Benchmark 1: cargo run --release -- --roux < tests/random/2
  Time (mean ± σ):     130.4 ms ±   5.0 ms    [User: 120.3 ms, System: 10.0 ms]
  Range (min … max):   122.1 ms … 143.5 ms    20 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 1: cargo run --release -- --cfop < tests/random/1
  Time (mean ± σ):      3.198 s ±  0.056 s    [User: 3.061 s, System: 0.138 s]
  Range (min … max):    3.132 s …  3.276 s    5 runs

Benchmark 1: cargo run --release -- --cfop < tests/random/2
  Time (mean ± σ):     320.6 ms ±   6.2 ms    [User: 294.7 ms, System: 25.8 ms]
  Range (min … max):   311.4 ms … 327.8 ms    5 runs
```

Add `cubes_from_goal` as visited map:

```
    Finished release [optimized] target(s) in 0.01s
Benchmark 1: cargo run --release -- --roux < tests/random/1
  Time (mean ± σ):     636.2 ms ±  30.1 ms    [User: 597.2 ms, System: 38.9 ms]
  Range (min … max):   603.7 ms … 702.5 ms    20 runs

Benchmark 1: cargo run --release -- --roux < tests/random/2
  Time (mean ± σ):     118.0 ms ±  13.2 ms    [User: 100.4 ms, System: 17.4 ms]
  Range (min … max):   109.5 ms … 173.1 ms    20 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 1: cargo run --release -- --cfop < tests/random/1
  Time (mean ± σ):      3.095 s ±  0.085 s    [User: 2.940 s, System: 0.155 s]
  Range (min … max):    3.024 s …  3.241 s    5 runs

Benchmark 1: cargo run --release -- --cfop < tests/random/2
  Time (mean ± σ):     271.3 ms ±   4.2 ms    [User: 255.7 ms, System: 15.7 ms]
  Range (min … max):   264.3 ms … 274.5 ms    5 runs
```

And add `cubes_from_start`:

```
   Compiling cube v0.5.2 (/home/cympfh/git/cube)
    Finished release [optimized] target(s) in 3.28s
Benchmark 1: cargo run --release -- --roux < tests/random/1
  Time (mean ± σ):     601.6 ms ±  26.9 ms    [User: 568.5 ms, System: 32.4 ms]
  Range (min … max):   572.7 ms … 668.7 ms    20 runs

Benchmark 1: cargo run --release -- --roux < tests/random/2
  Time (mean ± σ):     119.6 ms ±   5.9 ms    [User: 109.5 ms, System: 10.1 ms]
  Range (min … max):   110.0 ms … 128.1 ms    20 runs

Benchmark 1: cargo run --release -- --cfop < tests/random/1
  Time (mean ± σ):      3.623 s ±  0.040 s    [User: 3.473 s, System: 0.150 s]
  Range (min … max):    3.559 s …  3.662 s    5 runs

Benchmark 1: cargo run --release -- --cfop < tests/random/2
  Time (mean ± σ):     246.6 ms ±  17.8 ms    [User: 228.3 ms, System: 17.9 ms]
  Range (min … max):   222.9 ms … 267.6 ms    5 runs
```